Add all built XML strings
I've made xml strings that cover all system events according the the schema.
I think the way to do it is have a function that parses the string that
comes through the socket, then based on the type of event it calls a
different function to return the log string, passing it the rest of the data.

Ex.
audit('userCommand,<----comma delimited parameters---->')