Modularize client
This commit refactors the client to be a function call which
can be later edited to take a data string as a parameter.

This commit also contains some skeleton code on the server for
handling the requests and parsing them string->dict->xml

this function has yet to be run since I don't have access to a
dict containing the data in the form it is going to be passed.